tests/wpt/web-platform-tests/css/css-typed-om/the-stylepropertymap/inline/get-invalid.html

<!doctype html>
<meta charset="utf-8">
<title>Inline StylePropertyMap.get Error Handling</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#get-a-value-from-a-stylepropertymap">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../resources/testhelper.js"></script>
<body>
<div id="log">
<script>
'use strict';

test(t => {
  const styleMap = createInlineStyleMap(t);
  assert_throws_js(TypeError, () => styleMap.get('lemon'));
}, 'Calling StylePropertyMap.get with an unsupported property throws a TypeError');

</script>
